refactor: split rpy-parser.service.ts (2674 lines) into focused modules (#278)
Extract 7 focused modules from the monolithic rpy-parser service:
- rpy/types.ts — all interfaces and types
- rpy/helpers.ts — isValidLabel, trackBlocks, getIndent, countLinesInChoice
- rpy/technical-constructs.ts — splitConditionParts, extractTechnicalConstructs*
- rpy/parser.ts — extract* and parse* functions
- rpy/reconstruction.ts — reconstructRPYFile
- rpy/label-management.ts — label CRUD operations
- rpy/export-generator.ts — generateRpyFile, convertToBranchForgeFormatFromLabels

Original file becomes a barrel re-export for backward compatibility.
All consumer imports unchanged. Zero behavior changes.
